+++ e2e-tests/src/test/finalization.rs
@@ -1,7 +1,7 @@
 use common::create_connection;
 
 use crate::config::Config;
-use crate::utils::waiting::wait_for_finalized_block;
+use crate::waiting::wait_for_finalized_block;
 
 pub fn finalization(config: Config) -> anyhow::Result<u32> {
     let connection = create_connection(config.node);
