+++ e2e-tests/src/transfer.rs
@@ -4,9 +4,9 @@ use sp_core::Pair;
 use substrate_api_client::sp_runtime::AccountId32;
 use substrate_api_client::AccountId;
 
+use crate::accounts::accounts_from_seeds;
 use crate::config::Config;
-use crate::utils::accounts::accounts_from_seeds;
-use crate::utils::types::{Connection, TransferTransaction};
+use crate::{Connection, TransferTransaction};
 
 pub fn setup_for_transfer(config: Config) -> (Connection, AccountId32, AccountId32) {
     let Config { node, seeds, .. } = config;
