--- e2e-tests/src/utils/mod.rs
@@ -1,6 +1,0 @@
-pub mod accounts;
-pub mod extrinsic;
-pub mod fee;
-pub mod transfer;
-pub mod types;
-pub mod waiting;
